refactor: extract MCP/LSP init from app.tsx
Move initMcp and initLsp (~150 lines each) from app.tsx callback bodies
into src/app/mcp-lsp-init.ts as standalone async functions.

App keeps thin wrapper useCallbacks that delegate to the extracted
functions. All call sites remain unchanged — zero behavioral change.

Also removes now-unused imports (buildSystemMessages, isImagePath) from
app.tsx.

Commit-Queue-Session: cq_20260508160558_dad111b7260fef0bcd71f2b9
Coding-Agent: pi
Coding-Agent-Session: pi-019e06b6-830e-72ca-8378-e2ba2d3209c0
